requirements.txt
flask==1.1.2
sqlalchemy=1.3.23
marshmallow==3.10.0
pymysql==1.0.2
pyjwt==2.0.1
